refactor: Domain 루트 패키지 제거 및 내부 클래스들 각자 헥사고날 아키텍처에 맞는 위치로 이동, 기존 서비스 의존성 가지던 부분 수정
